extracting stringify function out of QueryResult component
